:truck: move model and admin files from api to core app
