+++ src/objects/conf/base.py
@@ -70,6 +70,7 @@ INSTALLED_APPS = [
     # Project applications.
     "objects.accounts",
     "objects.api",
+    "objects.core",
     "objects.utils",
 ]
